+++ src/HotChocolate/Core/src/Types.Mutations/MutationContextData.cs
@@ -4,7 +4,7 @@ namespace HotChocolate.Types;
 /// This internal data structure is used to store the effective mutation options of a field
 /// on the context so that the type interceptor can access them.
 /// </summary>
-internal class MutationContextData
+internal sealed class MutationContextData
 {
     public MutationContextData(
         ObjectFieldDefinition definition,
